requirements.txt
django-rest-swagger
django
djangorestframework
django-filter
